feat : suspense 확장을 위한 ui 분리
